requirements.txt
mcp>=0.9.0
aiohttp>=3.8.0
pandas>=1.5.0
beautifulsoup4>=4.11.0
google-ads>=20.0.0
tenacity>=8.0.0